docs: move data-grid pagination demos to examples
