+++ docs/examples/data-grid/pagination/change.demo.tsx
@@ -4,10 +4,10 @@ import { DataGrid } from '@cerberus-design/data-grid'
 import { useQuery } from '@cerberus-design/signals'
 import { Stack } from 'styled-system/jsx'
 import { queryEmployees } from '../api'
-import { columns } from '../quick-start/columns.demo'
+import { columns } from '../col-defs/employees'
 
-export function PageDemo() {
-  const data = useQuery(queryEmployees(1000))
+export function ChangeDemo() {
+  const data = useQuery(queryEmployees(100))
 
   return (
     <Stack direction="column" h="20rem" w="3/4">
@@ -15,7 +15,10 @@ export function PageDemo() {
         columns={columns}
         data={data}
         pagination={{
-          defaultPage: 2,
+          pageSize: 25,
+          onPageChange: (details) => {
+            console.log(details)
+          },
         }}
       />
     </Stack>
